explorer: fix memory-map overlap lookup and signal-bitmap width, balance signals macro
objectAtAddress no longer breaks early on the first miss — objects are sorted by
start address only, so an earlier, larger object can still cover the target.
Widen the signal bit map to int32 so messages with many signals can't wrap the
index negative. Wrap the protobuf json_util include in push/pop_macro(signals)
in the dbc/ldf sessions instead of leaving Qt's 'signals' macro undefined for
the rest of the TU. Drop a dead model query in SignalGridItem, generalize the
abseil DLL glob to any version, and fix the _color_map shade-encoding comment.

+++ src/models/signalmapmodel.cpp
@@ -318,7 +318,7 @@ void SignalMapModel::rebuildBitMap() {
         for (int pos : positions) {
             if (pos >= 0 && pos < totalBits) {
                 if (_bit_map[pos] < 0) {
-                    _bit_map[pos] = static_cast<int16_t>(si);
+                    _bit_map[pos] = si;
                 } else if (_bit_map[pos] != si) {
                     _overlap_map[pos] = true;
                 }
@@ -326,7 +326,7 @@ void SignalMapModel::rebuildBitMap() {
         }
     }
 
-    for (int16_t v : _bit_map) {
+    for (int32_t v : _bit_map) {
         if (v >= 0) ++_used_bits;
     }
 }
